backend/pyproject.toml
[project]
name = "nekkoflix-backend"
version = "0.1.0"
description = "nekkoflix backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "google-cloud-aiplatform>=1.71.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-storage>=2.19.0",
    "httpx>=0.28.0",
    "pillow>=11.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.6.0",
    "pre-commit>=4.5.1",
    "structlog>=24.4.0",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "N", "ANN"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"src/routers/*.py" = ["B008"]
"src/dependencies.py" = ["B008"]
"tests/**/*.py" = ["ANN"]

[tool.mypy]
python_version = "3.11"
strict = true
explicit_package_bases = true
ignore_missing_imports = false
disallow_untyped_defs = true
disallow_any_explicit = false
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "google.cloud.firestore",
    "google.cloud.firestore.*",
    "google.cloud.storage",
    "google.cloud.storage.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests", "tests.*"]
disallow_untyped_defs = false
disallow_untyped_calls = false
check_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
